Report exception only when last event indicates it
Retrieve the stored last event from the dump and use it to determine which thread had an exception. This is preferable to just checking `Thread.CurrentException` because 1) we will miss Win32 exceptions, and 2) the `CurrentException` property may contain residue from an exception that was previously handled and not yet cleared.