Add Release-NoCostura build configuration
This build configuration doesn't run Costura, which squishes all dependencies into the target binary. The native dependencies are also not embedded into the final binary, and are instead copied over to the output directory.

The rationale for adding this configuration is that the way Costura expands and loads managed and unmanaged dependencies can be problematic in some hosting environments, namely Azure App Service.