Set symbol path on native DbgEng target
Previously, the native DbgEng target used the default symbol path from the environment variable. When the environment variable is not set, this target would not have a symbol path at all, leading to symbol resolution failures. It turns out that setting the CLRMD symbol path is not enough, and we need to call `IDebugSymbols.SetSymbolPath` explicitly as well.